do sample prediction with `curl | ... |  base64`
